# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/period": {
      "post": {
        "operationId": "PeriodController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePeriodDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePeriodDto"
                }
              }
            }
          }
        },
        "tags": [
          "period"
        ]
      },
      "get": {
        "operationId": "PeriodController_findAll",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponsePeriodDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "period"
        ]
      }
    },
...
    "/period/{id}": {
      "get": {
        "operationId": "PeriodController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePeriodDto"
                }
              }
            }
          }
        },
        "tags": [
          "period"
        ]
      },
      "patch": {
        "operationId": "PeriodController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePeriodDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePeriodDto"
                }
              }
            }
          }
        },
        "tags": [
          "period"
        ]
      },
      "delete": {
        "operationId": "PeriodController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePeriodDto"
                }
              }
            }
          }
        },
        "tags": [
          "period"
        ]
      }
    },
    "/schedule": {
      "post": {
        "operationId": "ScheduleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseScheduleDto"
                }
              }
            }
          }
        },
        "tags": [
          "schedule"
        ]
      }
    },
    "/schedule/period/{id}": {
      "get": {
        "operationId": "ScheduleController_findAll",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sectionId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "subjectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "periodId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "semester",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "dayId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "classroomId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseScheduleDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "schedule"
        ]
      }
    },
    "/schedule/students/period/{id}": {
      "get": {
        "operationId": "ScheduleController_findAllStudents",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sectionId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "subjectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "periodId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "semester",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "dayId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "classroomId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseScheduleDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "schedule"
        ]
      }
    },
    "/schedule/{id}": {
      "get": {
        "operationId": "ScheduleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseScheduleDto"
                }
              }
            }
          }
        },
        "tags": [
          "schedule"
        ]
      },
      "patch": {
        "operationId": "ScheduleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseScheduleDto"
                }
              }
            }
          }
        },
        "tags": [
          "schedule"
        ]
      },
      "delete": {
        "operationId": "ScheduleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseScheduleDto"
                }
              }
            }
          }
        },
        "tags": [
          "schedule"
        ]
      }
    },
...
    "/section/period/{periodId}": {
      "get": {
        "operationId": "SectionController_findAll",
        "parameters": [
          {
            "name": "periodId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "subjectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "semester",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "dayId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseSectionDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "section"
        ]
      }
    },
...
    "/inscription/period/{id}": {
      "get": {
        "operationId": "InscriptionController_findAll",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sectionId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "careerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "schoolId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "subjectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "semester",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "schedules",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseInscriptionDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "inscription"
        ]
      }
    },
...
  },
  "info": {
    "title": "SIGEIT-API Documentation",
    "description": "API completa para el Sistema de Gestión de Información Educativa y Tecnológica",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "IdCreateEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "Teacher": {
        "type": "object",
        "properties": {}
      },
...
      "CreatePeriodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "type": "string"
          },
          "end": {
            "format": "date-time",
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "interval": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "copyPrevious": {
            "type": "boolean"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "start",
          "end",
          "startTime",
          "endTime",
          "stage",
          "interval",
          "duration",
          "copyPrevious",
          "status"
        ]
      },
      "ResponsePeriodDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "type": "string"
          },
          "end": {
            "format": "date-time",
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "interval": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "start",
          "end",
          "startTime",
          "endTime",
          "stage",
          "interval",
          "duration",
          "status"
        ]
      },
      "UpdatePeriodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "type": "string"
          },
          "end": {
            "format": "date-time",
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "interval": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "copyPrevious": {
            "type": "boolean"
          },
          "status": {
            "type": "boolean"
          }
        }
      },
      "CreateScheduleDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "classroom": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "day": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "section": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "period": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "start",
          "end",
          "classroom",
          "day",
          "section",
          "period",
          "status"
        ]
      },
      "Classroom": {
        "type": "object",
        "properties": {}
      },
      "Day": {
        "type": "object",
        "properties": {}
      },
      "Section": {
        "type": "object",
        "properties": {}
      },
      "Period": {
        "type": "object",
        "properties": {}
      },
      "ResponseScheduleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "classroomId": {
            "type": "number"
          },
          "classroom": {
            "$ref": "#/components/schemas/Classroom"
          },
          "dayId": {
            "type": "number"
          },
          "day": {
            "$ref": "#/components/schemas/Day"
          },
          "sectionId": {
            "type": "number"
          },
          "section": {
            "$ref": "#/components/schemas/Section"
          },
          "periodId": {
            "type": "number"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "start",
          "end",
          "classroomId",
          "classroom",
          "dayId",
          "day",
          "sectionId",
          "section",
          "periodId",
          "period",
          "status"
        ]
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "classroom": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "day": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "section": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "period": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          }
        }
      },
...
      "Subject": {
        "type": "object",
        "properties": {}
      },
      "ResponseSectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "capacity": {
            "type": "number"
          },
          "inscribed": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "subjectId": {
            "type": "number"
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          },
          "periodId": {
            "type": "number"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "teacherId": {
            "type": "number"
          },
          "teacher": {
            "$ref": "#/components/schemas/Teacher"
          },
          "status": {
            "type": "boolean"
          },
          "all": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "capacity",
          "inscribed",
          "name",
          "subjectId",
          "subject",
          "periodId",
          "period",
          "teacherId",
          "teacher",
          "status",
          "all"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {}
      },
      "ResponseInscriptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "stage": {
            "type": "string"
          },
          "sectionId": {
            "type": "number"
          },
          "section": {
            "$ref": "#/components/schemas/Section"
          },
          "userId": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "id",
          "stage",
          "sectionId",
          "section",
          "userId",
          "user"
        ]
      },
...
    }
  }
}